move scheduler to its own file
