+++ src/test/scala/chiseltest/tests/ThreadJoinTest.scala
@@ -2,8 +2,6 @@
 
 package chiseltest.tests
 
-import org.scalatest._
-
 import chisel3._
 import chiseltest._
 import org.scalatest.flatspec.AnyFlatSpec
